* Updated random monster stuffs -- Added db/[pre-]re/mob_random.txt -- Contains list of random monster files (mob_branch, mob_pouch, mob_boss, & mob_classchange) -- Restructured random monster format, now <mob group constant>,<mob id>,<rate> -- Added MOB_ constants for random monster group -- Max random monster each group, is not limited to 350 anymore * Move out some hardcodes to battle config file: -- Fame points for Taekwon mission, refine success, & pharmacy success (player.conf) -- Taekwon mission's monster criteria (skill.conf) * Monster db (mob_db and the import/mob_db.txt) now being read before mob_skill_db.txt to avoid skill being resetted when there is imported mob entry.

+++ src/map/pc.h
@@ -1036,7 +1036,7 @@ void pc_delinvincibletimer(struct map_session_data* sd);
 
 void pc_addspiritball(struct map_session_data *sd,int interval,int max);
 void pc_delspiritball(struct map_session_data *sd,int count,int type);
-void pc_addfame(struct map_session_data *sd,int count);
+void pc_addfame(struct map_session_data *sd, unsigned short count);
 unsigned char pc_famerank(int char_id, int job);
 bool pc_set_hate_mob(struct map_session_data *sd, int pos, struct block_list *bl);
